+++ components/Molecules/StaticHero/StaticHero.js
@@ -1,13 +1,13 @@
 import Image from 'next/image';
 
 const Hero = ({ image, alt, heroImages }) => (
-  <section>
+  <section data={heroImages}>
     <div className="hidden content md:block">
       <Image
         src={heroImages ? heroImages?.desktop : `/hero-${image}.jpg`}
         alt={`${alt} | CFC Capital`}
-        width={1280}
-        height={577}
+        width={1400}
+        height={635}
         style={{
           width: '100%',
           height: 'auto',
@@ -18,8 +18,8 @@ const Hero = ({ image, alt, heroImages }) => (
       <Image
         src={heroImages ? heroImages?.mobile : `/m-${image}.jpg`}
         alt={`${alt} | CFC Capital`}
-        width={1280}
-        height={577}
+        width={700}
+        height={500}
         style={{
           width: '100%',
           height: 'auto',