+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## [1.0.2](https://github.com/tavomak/Cfc-Capital-frontend/compare/v1.0.1...v1.0.2) (2024-11-28)
+
+### Bug Fixes
+
+- **design:** Hero images fix ([ffa0b13](https://github.com/tavomak/Cfc-Capital-frontend/commit/ffa0b139071a026206f6dce5f80d9348219568a5))
+
+## [1.0.1](https://github.com/tavomak/Cfc-Capital-frontend/compare/v1.0.0...v1.0.1) (2024-11-27)
+
+### Bug Fixes
+
+- **signature:** Empty phone ([2d50c32](https://github.com/tavomak/Cfc-Capital-frontend/commit/2d50c326cb2ff1bc8b511157b2228167fe86aacd))
+
 # 1.0.0 (2024-11-23)
 
 ### Features